Create package tree in hilt-processor module
