+++ hilt/processor/src/main/kotlin/poet/FactoryMethodsModulePoet.kt
@@ -15,7 +15,7 @@
  *
  */
 
-package it.czerwinski.android.hilt.processor
+package it.czerwinski.android.hilt.processor.poet
 
 import androidx.annotation.NonNull
 import com.squareup.javapoet.AnnotationSpec
@@ -27,6 +27,8 @@ import com.squareup.javapoet.TypeSpec
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
+import it.czerwinski.android.hilt.processor.model.FactoryMethodModel
+import it.czerwinski.android.hilt.processor.model.KotlinElementKind
 import java.util.Locale
 import javax.annotation.processing.Filer
 import javax.lang.model.element.Modifier
